requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.1.7
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.5.3
whitenoise==6.9.0
